Changed some (int) casts to (int)(long) to get rid of harmless warnings on 64 bit systems
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@20362 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/gui/c64ui.c
@@ -350,7 +350,7 @@ static UI_CALLBACK(save_screenshot)
 {
     /* Where does the 1024 come from?  */
     char filename[1024];
-    unsigned int wid = (unsigned int)UI_MENU_CB_PARAM;
+    unsigned int wid = (unsigned int)(unsigned long)UI_MENU_CB_PARAM;
 
     vsync_suspend_speed_eval();
