+++ vice/src/drive/iec/c64exp/c64exp-resources.c
@@ -55,24 +55,24 @@ static void set_drive_ram(unsigned int dnr)
 
 static int set_drive_parallel_cable(int val, void *param)
 {
-    drive_t *drive = drive_context[(unsigned int)param]->drive;
+    drive_t *drive = drive_context[(unsigned int)(unsigned long)param]->drive;
 
     if (val != DRIVE_PC_NONE && val != DRIVE_PC_STANDARD
         && val != DRIVE_PC_DD3)
         return -1;
 
     drive->parallel_cable = val;
-    set_drive_ram((unsigned int)param);
+    set_drive_ram((unsigned int)(unsigned long)param);
 
     return 0;
 }
 
 static int set_drive_profdos(int val, void *param)
 {
-    drive_t *drive = drive_context[(unsigned int)param]->drive;;
+    drive_t *drive = drive_context[(unsigned int)(unsigned long)param]->drive;;
 
     drive->profdos = val;
-    set_drive_ram((unsigned int)param);
+    set_drive_ram((unsigned int)(unsigned long)param);
 
     return 0;
 }
@@ -110,10 +110,10 @@ int c64exp_resources_init(void)
 
         res_drive[0].name = lib_msprintf("Drive%iParallelCable", dnr + 8);
         res_drive[0].value_ptr = &(drive->parallel_cable);
-        res_drive[0].param = (void *)dnr;
+        res_drive[0].param = (void *)(unsigned long)dnr;
         res_drive[1].name = lib_msprintf("Drive%iProfDOS", dnr + 8);
         res_drive[1].value_ptr = &(drive->profdos);
-        res_drive[1].param = (void *)dnr;
+        res_drive[1].param = (void *)(unsigned long)dnr;
 
         if (resources_register_int(res_drive) < 0)
             return -1;
